+++ src/types/index.ts
@@ -42,26 +42,6 @@ export type CurrentReport = {
   currentTerm: string;
 };
 
-export type UploadFile = {
-  id: string;
-  file: File;
-  [x: string]: string | File;
-};
-
-export type PresignedUrlResponse = {
-  id: string;
-  uploadUrl: string;
-  contentType: string;
-};
-
-export type UrlType = {
-  id?: string;
-  order?: number;
-  fileName?: string;
-  originUrl: string;
-  cdnUrl: string;
-};
-
 export type OrderUUID = { id: string; order: number };
 
 export type Auth = Pick<LoginResponse, 'role' | 'token'>;
